+++ src/httphandler/BaseHttpHandler.java
@@ -1,4 +1,4 @@
-package httpHandler;
+package httphandler;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
