+++ src/httphandler/EpicsHandler.java
@@ -1,4 +1,4 @@
-package httpHandler;
+package httphandler;
 
 import general.Const;
 import general.NotFoundException;
