+++ src/httphandler/HistoryHandler.java
@@ -1,4 +1,4 @@
-package httpHandler;
+package httphandler;
 
 import general.Const;
 import manager.HistoryManager;
