+++ src/httphandler/SubTasksHandler.java
@@ -1,29 +1,28 @@
-package httpHandler;
+package httphandler;
 
 import general.Const;
 import general.NotFoundException;
 import general.TaskAcrossException;
 import manager.TaskManager;
-import task.Task;
+import task.SubTask;
 
-public class TasksHandler extends BaseHttpHandler {
+public class SubTasksHandler extends BaseHttpHandler {
 
-    public TasksHandler(TaskManager taskManager) {
+    public SubTasksHandler(TaskManager taskManager) {
         super(taskManager);
     }
 
     @Override
     protected void post() {
-        Task task = gson.fromJson(jsonBody, Task.class);
+        SubTask subTask = gson.fromJson(jsonBody, SubTask.class);
         try {
-            if (task.getId() > 0) {
-                taskManager.updateTask(task);
+            if (subTask.getId() > 0) {
+                taskManager.updateSubTask(subTask);
             } else {
-                taskManager.addTask(task);
+                taskManager.addSubTask(subTask);
             }
             code = Const.CODE_SUCCESS_FOR_CREATE_UPD;
-        }
-        catch (TaskAcrossException e) {
+        } catch (TaskAcrossException e) {
             code = Const.CODE_ACCROSS_ERROR;
         }
     }
@@ -33,21 +32,20 @@ public class TasksHandler extends BaseHttpHandler {
         code = Const.CODE_SUCCESS;
         if (paramId > 0) {
             try {
-                response = gson.toJson(taskManager.getTask(paramId));
-            }
-            catch (NotFoundException e) {
+                response = gson.toJson(taskManager.getSubTask(paramId));
+            } catch (NotFoundException e) {
                 code = Const.CODE_NOT_FOUND;
                 response = "";
             }
         } else {
-            response = gson.toJson(taskManager.getAllTask());
+            response = gson.toJson(taskManager.getAllSubTask());
         }
     }
 
     @Override
     protected void delete() {
         if (paramId > 0) {
-            taskManager.delTask(paramId);
+            taskManager.delSubTask(paramId);
         }
         code = Const.CODE_SUCCESS;
     }
